Moved updated source code to src directory
